Move datasets for dispatch into their own package
