+++ internal/datasets/subjectsetbytype.go
@@ -1,4 +1,4 @@
-package util
+package datasets
 
 import (
 	"fmt"
