+++ internal/datasets/subjectsetbytype_test.go
@@ -1,4 +1,4 @@
-package util
+package datasets
 
 import (
 	"sort"
